점프 임시 구현 + 플레이어 : Component -> GameObject 하위 클래스로 변경 + TimeManager 클래스 생성 + Vec2 필드값 int->float 으로 변경

+++ Functions.cpp
@@ -6,7 +6,7 @@
 // ========== 문자열 처리 ==========
 wstring to_wstring(Vec2 vec)
 {
-	return std::to_wstring(vec.x) + L" , " + std::to_wstring(vec.y);
+	return std::to_wstring((int)vec.x) + L" , " + std::to_wstring((int)vec.y);
 }
 
 wstring to_wstring(const string& str)
